Add files via upload
In this 3 Transistors Audio Amplifier Circuit, first, the audio signal is fed into the circuit through mic 1. It is then converted into a tiny electrical signal that is sent to Q1. First, to amplify the signal we have configured a common emitter. Resistors  R1 and R2 are used to divide the voltage into bias to transistor Q1. However, the signal is insufficient. Therefore, it is transferred to transistors  Q2 and Q3, to amplify the signal so that it would be able to be driven into the speaker. 
Application Uses:
In mic, speaker audio applications.